+++ src/anyio/abc/_tasks.py
@@ -26,6 +26,20 @@ class TaskGroup(metaclass=ABCMeta):
         """
 
     @abstractmethod
+    async def start(self, func: Callable[..., Coroutine], *args, name=None) -> None:
+        """
+        Launch a new task and wait until it signals for readiness.
+
+        :param func: a coroutine function
+        :param args: positional arguments to call the function with
+        :param name: name of the task, for the purposes of introspection and debugging
+        :return: the value passed to ``task_status.started()``
+        :raises RuntimeError: if the task finishes without calling ``task_status.started()``
+
+        .. versionadded:: 3.0
+        """
+
+    @abstractmethod
     async def __aenter__(self) -> 'TaskGroup':
         """Enter the task group context and allow starting new tasks."""