+++ tests/test_taskgroups.py
@@ -80,6 +80,59 @@ async def test_spawn_after_error():
     exc.match('This task group is not active; no new tasks can be spawned')
 
 
+async def test_start_no_value():
+    async def taskfunc(*, task_status):
+        task_status.started()
+
+    async with create_task_group() as tg:
+        value = await tg.start(taskfunc)
+        assert value is None
+
+
+async def test_start_with_value():
+    async def taskfunc(*, task_status):
+        task_status.started('foo')
+
+    async with create_task_group() as tg:
+        value = await tg.start(taskfunc)
+        assert value == 'foo'
+
+
+async def test_start_crash_before_started_call():
+    async def taskfunc(*, task_status):
+        raise Exception('foo')
+
+    async with create_task_group() as tg:
+        with pytest.raises(Exception) as exc:
+            await tg.start(taskfunc)
+
+    exc.match('foo')
+
+
+async def test_start_crash_after_started_call():
+    async def taskfunc(*, task_status):
+        task_status.started(2)
+        raise Exception('foo')
+
+    with pytest.raises(Exception) as exc:
+        async with create_task_group() as tg:
+            value = await tg.start(taskfunc)
+
+    exc.match('foo')
+    assert value == 2
+
+
+async def test_start_no_started_call():
+    async def taskfunc(*, task_status):
+        pass
+
+    async with create_task_group() as tg:
+        with pytest.raises(RuntimeError) as exc:
+            await tg.start(taskfunc)
+
+    exc.match('hild exited')
+
+
 async def test_host_exception():
     async def set_result(value):
         nonlocal result
@@ -411,11 +464,13 @@ async def test_cancel_cascade():
         async with create_task_group() as tg2:
             tg2.spawn(sleep, 1)
 
+        print('raising exception')
         raise Exception('foo')
 
     async with create_task_group() as tg:
         tg.spawn(do_something)
         await wait_all_tasks_blocked()
+        print('all tasks blocking')
         tg.cancel_scope.cancel()
 
 
@@ -528,14 +583,16 @@ async def test_exception_group_filtering():
 async def test_cancel_propagation_with_inner_spawn():
     async def g():
         async with anyio.create_task_group() as tg:
+            print('inner task group:', id(tg))
             tg.spawn(anyio.sleep, 10)
-            await anyio.sleep(5)
+            await anyio.sleep(1)
 
         assert False
 
     async with anyio.create_task_group() as group:
+        print('outer task group:', id(group))
         group.spawn(g)
-        await anyio.sleep(0.1)
+        await wait_all_tasks_blocked()
         group.cancel_scope.cancel()